Refactor generate_test_failure_workbook.py to generate a zipfile
